+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 v33.0.0 (unreleased)
 --------------------
 
+- Remove the admin app and views.
+  https://github.com/nexB/scancode.io/issues/645
+
 v32.1.0 (2023-03-23)
 --------------------